requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.1
cssselect2==0.7.0
daphne==4.1.2
Django==5.0
django-htmx==1.19.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-phonenumber-field==8.0.0
django-shortuuidfield==0.1.3
django-tinymce==4.1.0
django-unfold==0.39.0
et-xmlfile==1.1.0
fonttools==4.54.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jsonfield==3.1.0
kombu==5.4.2
msgpack==1.1.0
openpyxl==3.1.5
packaging==24.1
phonenumbers==8.13.46
pillow==10.4.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyOpenSSL==24.2.1
pyphen==0.16.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
service-identity==24.1.0
setuptools==75.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
swapper==1.4.0
tinycss2==1.3.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.7.0
zope.interface==7.0.3
zopfli==0.2.3
